Version 2.0
Version 1.0 just wrote image URLs to a spreadsheet.  The links expired after a while so it became useless quickly. This version downloads the images and renames them with the episode title.  I also add a feature to print the loop number & episode title for better monitoring of the progress.  (I'm not sure it's creating every file.)